requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
scipy==1.10.1
tensorflow==2.12.0
flask==2.3.2
flask-socketio==5.3.4
plotly==5.14.1
dash==2.10.2
paho-mqtt==2.2.1
python-dotenv==1.0.0
pytest==7.3.1
